better eagle files
the traces are auto trash and there is a lot of "extra parts", but the form is coming together.  Width is close to the size of enclosures I am investigating.  The short dimension can still be shrunk further.  Currently this plan is about $18 at OshPark for three boards. That's OK.